NO-TICKET: refactoring
