+++ clients/mobile/__tests__/utility/formatting.test.ts
@@ -1,35 +1,35 @@
-import { formatteDate, formatPhoneNumber, formatErrorMessage } from '../../src/utility/formatte'
+import { formattedDate, formatPhoneNumber, formatErrorMessage } from '../../src/utility/formatting'
 
 describe('Utility Functions', () => {
-  describe('formatteDate', () => {
-    describe('formatteDate', () => {
+  describe('formattedDate', () => {
+    describe('formattedDate', () => {
       test('should format date string correctly', () => {
         const dateString = '2024-10-23T14:30:00Z' // This is in UTC
         const expected = '10/23/2024, 2:30 PM' // Expected output in UTC
-        expect(formatteDate(dateString)).toBe(expected)
+        expect(formattedDate(dateString)).toBe(expected)
       })
 
       test('should format Date object correctly', () => {
         const dateObject = new Date('2024-10-23T14:30:00Z') // This is in UTC
         const expected = '10/23/2024, 2:30 PM' // Expected output in UTC
-        expect(formatteDate(dateObject)).toBe(expected)
+        expect(formattedDate(dateObject)).toBe(expected)
       })
 
       test('should handle invalid date input', () => {
         const invalidDate = 'invalid-date-string'
         const expected = 'Invalid Date'
-        expect(formatteDate(invalidDate)).toBe(expected)
+        expect(formattedDate(invalidDate)).toBe(expected)
       })
 
       test('should handle empty string as input', () => {
         const expected = 'Invalid Date'
-        expect(formatteDate('')).toBe(expected)
+        expect(formattedDate('')).toBe(expected)
       })
     })
 
     test('should handle empty string as input', () => {
       const expected = 'Invalid Date'
-      expect(formatteDate('')).toBe(expected)
+      expect(formattedDate('')).toBe(expected)
     })
   })
