+++ clients/mobile/src/components/inputElement/DateTimePicker.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { Text, TouchableOpacity, View, StyleSheet } from 'react-native'
 import DateTimePicker from '@react-native-community/datetimepicker'
-import { formatteDate } from '../../utility/formatte'
+import { formattedDate } from '../../utility/formatting'
 import { useTheme } from '../../setup/theme/hooks/useTheme'
 import { Theme } from '../../setup/theme'
 import { commonStyles } from './commonStyles'
@@ -19,7 +19,6 @@ interface DateTimePickerComponentProps {
 const DateTimePickerComponent: React.FC<DateTimePickerComponentProps> = ({ label, value, onChange, showDatePicker, setShowDatePicker, error, isRequired = false }) => {
   const styles = createStyles(useTheme())
   const [isPickingTime, setIsPickingTime] = useState<boolean>(false)
-  console.log('value', value)
   const handleDateChange = (event: any, selectedDate: Date | undefined) => {
     if (event.type === 'dismissed') {
       setShowDatePicker(false)
@@ -62,7 +61,7 @@ const DateTimePickerComponent: React.FC<DateTimePickerComponentProps> = ({ label
         {isRequired && <Text style={styles.asterisk}>*</Text>}
       </Text>
       <TouchableOpacity onPress={() => { setShowDatePicker(true) }} style={styles.datePicker}>
-        <Text>{value !== null ? formatteDate(value) : 'Select Date & Time'}</Text>
+        <Text>{value !== null ? formattedDate(value) : 'Select Date & Time'}</Text>
       </TouchableOpacity>
       {showDatePicker && (
         <DateTimePicker
