+++ clients/mobile/src/donationWorkflow/createUpdateDonation/useBloodRequest.ts
@@ -6,7 +6,7 @@ import { createDonation, DonationResponse, updateDonation } from '../donationSer
 import { useNavigation, useRoute } from '@react-navigation/native'
 import { SCREENS } from '../../setup/constant/screens'
 import { DonationScreenNavigationProp, DonationScreenRouteProp } from '../../setup/navigation/navigationTypes'
-import { formatPhoneNumber } from '../../utility/formatte'
+import { formatPhoneNumber } from '../../utility/formatting'
 import { useFetchClient } from '../../setup/clients/useFetchClient'
 
 export const DONATION_DATE_TIME_INPUT_NAME = 'donationDateTime'
@@ -24,7 +24,7 @@ export interface BloodRequestData {
   transportationInfo?: string;
 }
 
-interface BloodRequestDataErrors extends Omit<BloodRequestData, 'patientName' | 'shortDescription' | 'transportationInfo'> {}
+interface BloodRequestDataErrors extends Omit<BloodRequestData, 'patientName' | 'shortDescription' | 'transportationInfo'> { }
 
 const validationRules: Record<keyof BloodRequestDataErrors, ValidationRule[]> = {
   urgencyLevel: [validateRequired],
